major refactor coordinates added environment source

+++ Writer/Extensions/ColorExtensions.cs
@@ -5,7 +5,7 @@ namespace WhiteMagic.PanelClock
 {
     public static class ColorExtensions
     {
-        public static Color Scale(this Color c, double scale)
+        public static Color Scale(this Color c, float scale)
         {
             var r = (byte)Math.Min(255, c.R * scale);
             var g = (byte)Math.Min(255, c.G * scale);
@@ -13,7 +13,7 @@ namespace WhiteMagic.PanelClock
             return Color.FromArgb(c.A, r, g, b);
         }
 
-        public static Color ScaleAlpha(this Color c, double scale)
+        public static Color ScaleAlpha(this Color c, float scale)
         {
             var a = (byte)Math.Min(255, c.A * scale);
             return Color.FromArgb(a, c.R, c.G, c.B);
